Bring User Guide files into the repository (#11)
The GitHub wiki format has several limitations.
Bringing the files into the repository means that updates to documentation can merged at the same time as corresponding updates to code.
The files were copied from the markdowns used to produce the wiki, and our deploy process in ant now includes a target to render the markdowns into html. This is accomplished using the mkdocs tool. Mkdocs makes a very professional looking site, and it is compatible with GitHub Pages and Read the Docs. The changes on this branch include all of the configuration to allow both sites to host the rendered User Guide.